After comparing results of algorithm before last meeting with Guido and todays version, this version works better on testseq3. Also, reducing subpixel factor from 1000 to 100 makes virtually no diffrence - big question is why! But it doesn't and removal of one decimal enables me to reduce flow_x and flow_y to int16, avoiding problems with variable overflow further down paparazzi program. Maybe this even solves problem with program crashing on the drone.

+++ image.h
@@ -59,8 +59,8 @@ struct point_t {
 /* Vector structure for point differences */
 struct flow_t {
   struct point_t pos;         ///< The original position the flow comes from
-  int32_t flow_x;             ///< The x direction flow in subpixels // CHANGED 16 -> 32
-  int32_t flow_y;             ///< The y direction flow in subpixels // CHANGED 16 -> 32
+  int16_t flow_x;             ///< The x direction flow in subpixels // CHANGED 16 -> 32
+  int16_t flow_y;             ///< The y direction flow in subpixels // CHANGED 16 -> 32
 };
 
 /* Usefull image functions */
